==> notebooks/09_findings.ipynb <==
# Findings

# === SETUP ===

# standard libraries
from pathlib import Path
import pandas as pd
import numpy as np
import re
from tqdm import tqdm
import json

# for plotting
import matplotlib.pyplot as plt
import seaborn as sns

# paths
REPO_DIR = Path(".").resolve().parents[0]
DATA = REPO_DIR / "data"
FINDINGS = REPO_DIR / "results" / "final_findings"
FIGURES = FINDINGS / "figures"
FINDINGS.mkdir(parents=True, exist_ok=True)
FIGURES.mkdir(parents=True, exist_ok=True)
DEBATES = DATA / "debates_full.csv"
MEDIA = DATA / "media_balanced.csv"

# color palette
with open(Path(REPO_DIR / "color_palette_config.json")) as f:
    palette = json.load(f)

# load data
debates = pd.read_csv(DEBATES)
media = pd.read_csv(MEDIA)
